CREATE-3664: coverage xml should be updated only if the pull request is merged into develop

+++ batch_scripts/test_cov_analysis.py
@@ -6,8 +6,20 @@ import xml.etree.ElementTree as ET
 coverage_home_folder = sys.argv[1]
 print(coverage_home_folder)
 coverage_backup_path = f'{coverage_home_folder}coverage_report_backup.xml'
+print(coverage_backup_path)
 coverage_old_xml_path = f'{coverage_home_folder}coverage_report.xml'
+print(coverage_old_xml_path)
 coverage_new_xml_path = 'coverage_report.xml'
+print(coverage_new_xml_path)
+
+path = os.getcwd() 
+print("Current Directory", path) 
+
+parent_folder = os.path.abspath(os.path.join(path, os.pardir))
+print("Parent folder", parent_folder)
+
+is_develop = "develop" in parent_folder
+print(is_develop)
 
 coverage_old_xml = ET.parse(coverage_old_xml_path)
 coverage_new_xml = ET.parse(coverage_new_xml_path)
@@ -19,9 +31,10 @@ if coverage_new.attrib['line-rate'] < coverage_old.attrib['line-rate']:
     raise Exception("Test coverage is decreasing! You have to write unit tests to cover your new code lines")
 else:
     print("Test coverage is sufficient")
-    try:
-        os.remove(coverage_backup_path)
-        shutil.move(coverage_old_xml_path, coverage_backup_path)
-        shutil.copyfile(coverage_new_xml_path, coverage_old_xml_path)    
-    except IOError:
-        print("One of the coverage files is not accessible")
+    if is_develop:
+        try:
+            os.remove(coverage_backup_path)
+            shutil.move(coverage_old_xml_path, coverage_backup_path)
+            shutil.copyfile(coverage_new_xml_path, coverage_old_xml_path)    
+        except IOError:
+            print("One of the coverage files is not accessible")
